\name{Bayesiantreg-package}
\alias{Bayesiantreg-package}
\docType{package}
\title{Function to do Bayesian t Regression: joint mean and variance modeling and estimation of the degrees of freedom}
\description{Bayesian t regression package}
\details{
\tabular{ll}{
Package: \tab Bayesiantreg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2020-05-31\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
 Margarita Marin \email{mmarinj@unal.edu.co},
 Edilberto Cepeda-Cuervo \email{ecepedac@unal.edu.co}        
}
\keyword{package}




