\name{dJ2}
\alias{dJ2}
\title{density of the Jeffrey's distribution}
\description{calculates the density of the Jeffrey's distribution}
\usage{
dJ2(gl.ini, p)
}
\arguments{
  \item{gl.ini}{a vector with the number to evaluate in the density.}
  \item{p}{a number indicating the parameter value for the Jeffrey's prior an the Jeffrey's proposal}
}
\details{
Calculates the density of the Jeffrey's distribution
}
\value{
\item{J1}{the value of the density}
}
\author{
 Margarita Marin \email{mmarinj@unal.edu.co},
 Edilberto Cepeda-Cuervo \email{ecepedac@unal.edu.co}        
}
\keyword{Bayesian}
\keyword{Metropolis Hastings}
\keyword{Meancovariance modelling}
\keyword{t}
\keyword{Degrees of Freedom}
