\name{rJ2}
\alias{rJ2}
\title{random number from the Jeffrey's distribution}
\description{generates random numbers from the Jeffrey's distribution}
\usage{
rJ2(n, matriz, min, max)
}
\arguments{
  \item{n}{a number that indicates the number of random values that will by generate from the Jeffrey's distribution.}
  \item{matriz}{a matrix generate by the function tabla of the bayesiantreg package.}
  \item{min}{a number indicatein the minimum number that can be generated from the Jeffrey's distribution.}
  \item{max}{a number indicatein the maximum number that can be generated from the Jeffrey's distribution.}
}
\details{
generates random numbers from the Jeffrey's distribution
}
\value{
\item{grados}{the random number}
}
\author{
 Margarita Marin \email{mmarinj@unal.edu.co},
 Edilberto Cepeda-Cuervo \email{ecepedac@unal.edu.co}        
}
\keyword{Bayesian}
\keyword{Metropolis Hastings}
\keyword{Meancovariance modelling}
\keyword{t}
\keyword{Degrees of Freedom}
