\name{tabla}
\alias{tabla}
\title{probabilities and numbers from the Jeffrey's distribution}
\description{generates a table with diferente probabilities and associated numbers from the Jeffrey's distribution}
\usage{
tabla(min, max, p)
}
\arguments{
  \item{min}{a number indicatein the minimum number that can be generated from the Jeffrey's distribution.}
  \item{max}{a number indicatein the maximum number that can be generated from the Jeffrey's distribution.}
  \item{p}{a number indicating the parameter value for the Jeffrey's prior an the Jeffrey's proposal.}
}
\details{
generates a table with diferente probabilities and associated numbers from the Jeffrey's distribution
}
\value{
\item{matriz}{a matrix with the generated probabilities and associated numbers from the Jeffrey's distribution}
}
\author{
 Margarita Marin \email{mmarinj@unal.edu.co},
 Edilberto Cepeda-Cuervo \email{ecepedac@unal.edu.co}        
}
\keyword{Bayesian}
\keyword{Metropolis Hastings}
\keyword{Meancovariance modelling}
\keyword{t}
\keyword{Degrees of Freedom}
